+++ cmd/neofs-cli/modules/netmap/snapshot.go
@@ -6,6 +6,7 @@ import (
 	"github.com/nspcc-dev/neofs-node/cmd/internal/cmdprinter"
 	internalclient "github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/client"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/commonflags"
+	"github.com/nspcc-dev/neofs-sdk-go/client"
 	"github.com/spf13/cobra"
 )
 
@@ -24,15 +25,12 @@ var snapshotCmd = &cobra.Command{
 		}
 		defer cli.Close()
 
-		var prm internalclient.NetMapSnapshotPrm
-		prm.SetClient(cli)
-
-		res, err := internalclient.NetMapSnapshot(ctx, prm)
+		res, err := cli.NetMapSnapshot(ctx, client.PrmNetMapSnapshot{})
 		if err != nil {
 			return fmt.Errorf("rpc error: %w", err)
 		}
 
-		cmdprinter.PrettyPrintNetMap(cmd, res.NetMap())
+		cmdprinter.PrettyPrintNetMap(cmd, res)
 		return nil
 	},
 }